refactor(dexpi): split to_procedural into the read/procedural package
read/procedural/ holds the importer by concern: reports.py (ImportIssue,
DexpiImportReport, dexpi_import_report), resolve.py (_Endpoint, _Index,
_SegmentSpec, segment and signal spec resolution, endpoint resolution),
synthesize.py (equipment names/items/metadata and the junction,
instrument and inline synthesis), layout.py (layout rules, connectivity
grouping, site terminals, base-placement restore) and merge.py
(_fold_branch_groups, _merge_systems); __init__.py composes
dexpi_to_resolved, resolved_to_procedural_doc and dexpi_to_procedural_doc
with the same signatures. Every function and class body was lifted
verbatim by name; no name changed. read/to_procedural.py stays as an
explicit re-export facade of every former module-level name, so callers
and tests import as before. The one test that monkeypatched plan_layout
on the old module now patches the package that makes the call, since a
patch on the facade would no longer intercept anything.
